EAEC-45: Add SITE_URL for sitemap gen in docker & platform environments
Note 1: the next_dev environment does not currently utilize this variable because
sitemap is only built in the next_prod build task.

Note 2:  the prod url is hardcoded in platform.app.yaml because we
dont need to be pointing people to the development environments which
should be password protected anyway.